export graph to turtle file
