+++ src/features/edition/Project.jsx
@@ -1,8 +1,9 @@
-import { CollectionsBookmark, Edit } from '@mui/icons-material'
-import { ListItem, ListItemIcon, ListItemText, Typography, Button, IconButton } from '@mui/material'
+/* eslint-disable react-hooks/exhaustive-deps */
+import { CollectionsBookmark, Downloading, Edit } from '@mui/icons-material'
+import { ListItem, ListItemIcon, ListItemText, Typography, IconButton, Tooltip, CircularProgress } from '@mui/material'
 import { TimelineDot, TimelineSeparator, TimelineConnector } from '@mui/lab'
 import { Stack } from '@mui/system'
-import { useGetAnalyticalProjectQuery, useGetAnnotationsQuery } from '../../services/sparql'
+import { useExportProjectQuery, useGetAnalyticalProjectQuery, useGetAnnotationsQuery } from '../../services/sparql'
 import { useSelector } from 'react-redux'
 import { useEffect, useState } from 'react'
 import { Annotation } from './Annotation'
@@ -14,6 +15,8 @@ export const Project = () => {
   const [annotationsByPage, setAnnotationsByPage] = useState([])
   const { data: project } = useGetAnalyticalProjectQuery(projectIri, { skip: !projectIri })
   const { data: annotations } = useGetAnnotationsQuery({ scoreIri, projectIri }, { skip: !projectIri })
+  const [isDownloading, setIsDownloading] = useState(false)
+  const { data } = useExportProjectQuery('yes', { skip: !isDownloading })
 
   useEffect(() => {
     if (annotations)
@@ -27,6 +30,24 @@ export const Project = () => {
       )
   }, [annotations])
 
+  const downloadFile = async () => {
+    try {
+      const blob = new Blob([data], { type: 'text/turtle' })
+      const file = new File([blob], project.label + '.ttl', { type: 'text/turtle' })
+      const downloadLink = document.createElement('a')
+      downloadLink.href = URL.createObjectURL(file)
+      downloadLink.download = project.label + '.ttl'
+      downloadLink.click()
+      setIsDownloading(false)
+    } catch (e) {
+      console.error(e)
+    }
+  }
+
+  useEffect(() => {
+    if (data && isDownloading) downloadFile()
+  }, [data])
+
   if (project && annotations)
     return (
       <Stack flex={1} borderRadius={3} bgcolor="white" boxShadow={1} minHeight={0}>
@@ -36,9 +57,11 @@ export const Project = () => {
             <ListItem
               dense
               secondaryAction={
-                <Button size="small" disabled>
-                  publish
-                </Button>
+                <Tooltip title="Download Turtle file">
+                  <IconButton color="inherit" onClick={() => setIsDownloading(true)} edge="end">
+                    {isDownloading ? <CircularProgress color="inherit" size={20} /> : <Downloading />}
+                  </IconButton>
+                </Tooltip>
               }
             >
               <ListItemIcon>
