+++ src/services/sparql.js
@@ -3,15 +3,7 @@ import { getPage, stringToColor } from '../utils'
 //import { getContributor } from 'sherlock-sparql-queries/src/queries/contributor'
 //import { getAnalyticalProject } from 'sherlock-sparql-queries/src/queries/analyticalProject'
 import { DEV_ENV } from '../config/services'
-import {
-  exportEntity,
-  getAnalyticalProject,
-  getAnnotations,
-  getAssignments,
-  getContributor,
-  getP140,
-  getProjects,
-} from './queries'
+import { getAnalyticalProject, getAnnotations, getAssignments, getContributor, getP140, getProjects } from './queries'
 
 const SPARQL_ENDPOINT = DEV_ENV ? 'http://localhost:3030/iremus' : 'https://sherlock.freeboxos.fr/sparql'
 
@@ -79,17 +71,17 @@ export const sparql = createApi({
           page: getPage(page.value),
         })),
     }),
-    exportEntity: builder.query({
-      query: entity => ({
-        method: 'POST',
-        body: new URLSearchParams({ query: exportEntity(entity) }),
+    exportProject: builder.query({
+      query: () => ({
+        url:
+          '?' +
+          new URLSearchParams({
+            graph: 'http://data-iremus.huma-num.fr/graph/sherlock',
+          }),
+        method: 'GET',
+        headers: { 'Content-Type': 'text/turtle' },
+        responseHandler: response => response.text(),
       }),
-      transformResponse: response =>
-        response.results.bindings.map(({ s, p, o }) => ({
-          subject: s.value,
-          predicate: p.value,
-          object: o.value,
-        })),
     }),
     getProjects: builder.query({
       query: scoreIri => ({
@@ -115,5 +107,5 @@ export const {
   useGetP140Query,
   useGetAssignmentsQuery,
   useGetProjectsQuery,
-  useExportEntityQuery,
+  useExportProjectQuery,
 } = sparql
